requirements.txt
amqp==5.3.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
billiard==4.2.1
boto3==1.40.32
botocore==1.40.32
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
cssselect2==0.8.0
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
django-filter==25.1
django-ninja==1.4.3
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fonttools==4.59.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.1
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
PyMySQL==1.1.2
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
